Correct Reference Genomes
The ones I had from the file didn’t match, I went through and selected
the top peaks based on CCS data and mapped these back to the references
I received from Dave.  Several showed the original reference was
partially unrolled, so were trimmed. But the All4mer sequence was
rather off.  I took the consensus of the top 5 reads with the most
coverage as the “reference” for this case.