Fix ErrorReporter for Lambda
There was a bug that caused a random 50% of the lambda reads to be
skipped.  In particular, no results were output if no subreads were
found, but the subreads were matched to the screening templates ZMW, so
it was a crap shoot.  Also, I changed the ErrorReporter to not output
“no errors” in the event of no alignment.