Add emission probability counterweight
A deletion move inherently involves fewer steps than an emission move
because the deletion only has a transition probability associated with
it, while the emission move has not only a transition, but also up to 2
emission events associated with it.  This discrepancy can lead to
deletions having higher probabilities than matches which can goof the
banding optimizations, as they can greedily decide to “delete move” to
the upper right corner of the recursion matrix and emit barely (or not
at all) for the first few steps.  Unfortunately, this results in
globally suboptimal behavior as it favors paths or all delete followed
by all insert, which leads to underflow.  To counterbalance the
emission probability mismatch and encourage moving along the diagonal,
this commit introduces a scaling factor for emission events to
encourage them early on in the dynamic programming and counterbalance
the deletion > match behavior in the two-emission model.